Refactor renderTodoList.js and Fix renderModal.js
- separate rendering of different sections into there own file under
renderDomElements folder
- Fix renderModal.js to check if a container DOM element exist before
creating one and remove it
